requirements.txt
flask
flask-socketio
python-socketio[client]
eventlet
mss
Pillow
psutil
pynput
pygetwindow
requests

socketio
base64
io
socket
os
subprocess
threading
time
json
psutil
ctypes
webbrowser